renamed re-color mesh to recolor mesh
Hopefully this will make it easier for people to find the component
